+++ src/components/Editor/Runners/PythonRunner/PythonRunner.test.js
@@ -4,8 +4,12 @@ import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
 
 import PythonRunner from "./PythonRunner";
-import { codeRunHandled, setError, triggerDraw } from "../../EditorSlice";
-import { SettingsContext } from "../../../../settings";
+import {
+  codeRunHandled,
+  setError,
+  triggerDraw,
+} from "../../../../redux/EditorSlice";
+import { SettingsContext } from "../../../../utils/settings";
 import { matchMedia, setMedia } from "mock-match-media";
 import { MOBILE_BREAKPOINT } from "../../../../utils/mediaQueryBreakpoints";