+++ src/components/RunButton/RunButton.jsx
@@ -1,0 +1,26 @@
+import Button from "../Button/Button";
+
+import React from "react";
+import { useDispatch } from "react-redux";
+import { triggerCodeRun } from "../../redux/EditorSlice";
+
+const RunButton = ({ embedded = false, className, ...props }) => {
+  const dispatch = useDispatch();
+
+  const onClickRun = () => {
+    if (window.plausible) {
+      window.plausible(`Run button${embedded ? " embedded" : ""}`);
+    }
+    dispatch(triggerCodeRun());
+  };
+
+  return (
+    <Button
+      className={`btn--primary btn--run${className ? ` ${className}` : ""}`}
+      onClickHandler={onClickRun}
+      {...props}
+    />
+  );
+};
+
+export default RunButton;